Added sensors to QEV2 & restructured filesystem ... BUT CAMERA IS CAUSING COLLISION.
The added camera, despite floating distance atop of chassis, is causing the entire car to _dance_ uncontrollably.

FIX REQUIRED.

Other changes:

.git is now moved one folder up to include our custom models, i.e. LiDAR and camera.
This caused a rename in original filestructure but contents are generally unchanged.

The submodule has been removed. And since it is a catkin_ws, which is not very relevant to this repo, I've decided to keep it out.

The catkin repo is still accessible from QUTMS's QUTMS_catkin_ws repo.